requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiolimiter==1.2.1
aiosignal==1.3.2
aiosqlite==0.20.0
aiowpi==0.1.7
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==24.3.0
black==24.10.0
Brotli==1.1.0
cffi==1.17.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
loguru==0.7.3
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nonebot-adapter-onebot==2.4.6
nonebot-plugin-apscheduler==0.5.0
nonebot-plugin-waiter==0.8.0
nonebot2==2.4.1
numpy==2.2.2
opencv-python-headless==4.11.0.86
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pygtrie==2.5.0
pypika-tortoise==0.3.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.41.3
tortoise-orm==0.23.0
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
yarl==1.18.3
